requirements.txt
asgiref == 3.5.2
python_version >= "3.8"
django - cors - headers == 3.13.0
python_version >= "3.7"
django == 4.0.5
djangorestframework == 3.13.1
python_version >= "3.6"
pymysql == 1.0.2
python_version >= "3.6"
pytz == 2022.1
python_version >= "3.6"
sqlparse == 0.4.2
python_version >= "3.8"
tzdata == 2022.1
sys_platform == "win32" and python_version >= "3.8"
